Assert fixture before running expectation and refactor
