+++ projects/ngx-testing-extra-demo/src/app/components/page-2/page-2.component.spec.ts
@@ -2,13 +2,13 @@ import { componentTestBed } from 'ngx-testing-extra';
 import { Page2Component } from './page-2.component';
 
 describe('Page2Component', () => {
-  const bed = componentTestBed(Page2Component);
+  const tb = componentTestBed(Page2Component);
 
-  beforeEach(() => bed.compile());
+  beforeEach(() => tb.compile());
 
-  bed.shouldCreate();
+  tb.shouldCreate()
 
-  it('should click', bed(({ component, action }) => {
+  it('should click', tb(({ component, action }) => {
     expect(component.clicked).toBeFalse();
 
     action.click('#my-span');
