+++ projects/ngx-testing-extra/src/lib/components/element/cannot-find.ts
@@ -1,13 +1,13 @@
 import { Type } from '@angular/core';
 
-export function throwCannotFind(selectorOrDirective: string | Type<any>, many: boolean = false): never {
+type Quantifier =
+  | 'one'
+  | 'many'
+
+export function throwCannotFind(selectorOrDirective: string | Type<any>, quantifier: Quantifier = 'one'): never {
   const input: string = (typeof selectorOrDirective === 'string')
     ? `selector "${selectorOrDirective}"`
     : `directive "${selectorOrDirective.name}"`;
-
-  const quantifier: string = (many)
-    ? 'many'
-    : 'one';
 
   throw new Error(`Cannot find ${quantifier} DebugElement with : ${input}.`);
 }
