+++ projects/ngx-testing-extra/src/lib/components/test-bed/models/assert-fixture.ts
@@ -1,0 +1,4 @@
+export function assertComponentFixture(fixture: unknown): void {
+  if (!fixture)
+    throw new Error('ComponentFixture is falsy. You need to use `testBed.compile()` before running expectation.');
+}
