+++ projects/ngx-testing-extra/src/tests/units/components/element/cannot-find.spec.ts
@@ -1,4 +1,4 @@
-import { throwCannotFind } from '../../../../lib/components/element/debug-error';
+import { throwCannotFind } from '../../../../lib/components/element/cannot-find';
 import { NoWhereDirective } from '../../../fixtures/directives/no-where.directive';
 
 describe('throwCannotFind', () => {
@@ -7,7 +7,7 @@ describe('throwCannotFind', () => {
     expect(() => throwCannotFind('button'))
       .toThrowError('Cannot find one DebugElement with : selector "button".');
 
-    expect(() => throwCannotFind(NoWhereDirective, true))
+    expect(() => throwCannotFind(NoWhereDirective, 'many'))
       .toThrowError('Cannot find many DebugElement with : directive "NoWhereDirective".');
   });
 });
